Allow `no_std` and fewer dependencies
We need to be able to use this crate in a `no_std` context.
Additionally, we cannot make use of the `sha2` or `sha3` crates.
This commit adds a couple of features:

`std`, which is enabled by default, must be specified in order to use
`write_keypair_to_file`.  This was actually the only real dependency on
`std` outside of tests.

I've split the `bbsplus` feature up. `min_bbs` enables the algorithms,
but does not actually implement either the Sha256 or Shake256 schemes.

I've also bumped group to 0.13, which should be nonbreaking because
bls12_381_plus 0.8.18 already requires it.